+++ demo_glium/src/main.rs
@@ -3,8 +3,6 @@
 #![warn(clippy::all)]
 
 fn main() {
-    let title = "Egui glium demo";
-
     // Persist app state to file:
     let storage = egui_glium::storage::FileStorage::from_path(".egui_demo_glium.json");
 
@@ -12,5 +10,5 @@ fn main() {
     // let storage = egui::app::DummyStorage::default();
 
     let app: egui::DemoApp = egui::app::get_value(&storage, egui::app::APP_KEY).unwrap_or_default();
-    egui_glium::run(title, Box::new(storage), Box::new(app));
+    egui_glium::run(Box::new(storage), Box::new(app));
 }