+++ example_glium/src/main.rs
@@ -7,8 +7,6 @@ mod example_app;
 use example_app::ExampleApp;
 
 fn main() {
-    let title = "My Egui Window";
-
     // Persist app state to file:
     let storage = egui_glium::storage::FileStorage::from_path(".egui_example_glium.json");
 
@@ -18,5 +16,5 @@ fn main() {
     // Restore `example_app` from file, or create new `ExampleApp`:
     let app: ExampleApp = egui::app::get_value(&storage, egui::app::APP_KEY).unwrap_or_default();
 
-    egui_glium::run(title, Box::new(storage), Box::new(app));
+    egui_glium::run(Box::new(storage), Box::new(app));
 }